PathUtil stuff

+++ Program.fs
@@ -7,25 +7,25 @@ module Application =
     open Chiron
     open System
     open System.IO
-    open Dotnet2Nix.Util.JsonUtil
+    open Dotnet2Nix.Util
 
     let rec loadLibraries (file:string) =
         let data = File.ReadAllText(file)
         let json = Json.parse data
 
         let libs =
-            match getkey "libraries" json with
+            match JsonUtil.getkey "libraries" json with
               | Some (Json.Object o) -> o
               | _ -> raise (new Exception("Library json was not an object"))
 
         let combineProjectLibs libs _ lib =
             let libPath = 
               lazy 
-              match strkey "path" lib with
+              match JsonUtil.strkey "path" lib with
                 | Some s -> s
                 | _ -> raise (new Exception("No path found in project"))
 
-            match strkey "type" lib with
+            match JsonUtil.strkey "type" lib with
               | Some "project" ->
                   let otherProj = Path.GetDirectoryName(libPath.Force())
                   try 
@@ -46,26 +46,32 @@ module Application =
             | _ -> raise (new Exception("name and version not found in library name"))
 
         let sha512 = 
-          getkeyf "sha512" lib gets
+          JsonUtil.getkeyf "sha512" lib JsonUtil.gets
             |> Convert.FromBase64String 
-            |> byteHexStr
+            |> JsonUtil.byteHexStr
 
         let path = 
-          getkeyf "path" lib gets
+          JsonUtil.getkeyf "path" lib JsonUtil.gets
 
         let filterFile (file:string) =
           not (file.EndsWith(".nuspec") || file.EndsWith(".txt"))
 
+        // TODO: filter long 
         let outputFiles = 
-          getkeyf "files" lib geta
-            |> List.filter (gets >> Option.map filterFile >> defaultRaise "file was not a string")
+          JsonUtil.getkeyf "files" lib JsonUtil.geta
+            |> List.map (JsonUtil.gets >> JsonUtil.defaultRaise "file was not a string")
+            |> List.filter filterFile
+
+        // need to do this because of bash: argument list too long
+        let longestCommon =
+          PathUtil.longestCommonPath (List.toArray outputFiles)
 
         let (objMap:Map<string, Json>) = 
             Map.ofList [ "baseName", Json.String name
                          "version", Json.String ver 
                          "sha512", Json.String sha512
                          "path", Json.String path
-                         "files", Json.Array outputFiles
+                         "files", Json.Array (List.map Json.String outputFiles)
                        ]
 
         Json.Object objMap :: libs
@@ -75,16 +81,24 @@ module Application =
 
     [<EntryPoint>]
     let main argv =
-        let input = Array.tryItem 0 argv
-        let output = Array.tryItem 1 argv 
-                       |> Option.fold (fun _ file -> file) "nuget-packages.json"
-        match input with
-          | None -> usage.Force()
-          | Some filename ->
-              let libs = loadLibraries filename
-              let pkgs = Map.fold makePackage List.empty libs
-              let serialized = Json.formatWith JsonFormattingOptions.Pretty (Json.Array pkgs)
-              Console.Error.Write(String.Format("writing to {0}", output))
-              File.WriteAllText(output, serialized)
-              0
+
+        let input = 
+          Array.tryItem 0 argv
+            |> Option.fold (fun _ f -> f) (Path.Combine("obj", "project.assets.json"))
+
+        let output = 
+          Array.tryItem 1 argv 
+            |> Option.fold (fun _ file -> file) "nuget-packages.json"
+
+        let inputExists = File.Exists(input)
+
+        if not inputExists
+          then usage.Force()
+          else
+            let libs = loadLibraries input
+            let pkgs = Map.fold makePackage List.empty libs
+            let serialized = Json.formatWith JsonFormattingOptions.Pretty (Json.Array pkgs)
+            Console.Error.Write(String.Format("writing to {0}", output))
+            File.WriteAllText(output, serialized)
+            0
